Changes to SM and project save/open
Project open is corrected to allow cells arrays of any type (before only of structures). Created the read h5 functions to allow opening the h5 files, and supporting 1D arrays of cells.

Updates in the save and open sweep functions, now called export and import, to read data in the correct form and update the project accordingly.